By considering the terms in the Fibonacci sequence whose values do not exceed 4,000,000, write a program that finds and prints the sum of the even-valued terms, followed by a new line.

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -13,7 +13,7 @@ int main(void)
 		{
 			sum += k;
 		}
-		long temp = k;
+		int temp = k;
 		k = j + k;
 		j = temp;
 	}
